requirements.txt
PyQt6
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
